Moved the ColorLut code into its own folder
