+++ code/core/arm9/source/GbaColorCorrection/ColorLut.cpp
